requirements.txt
python-dotenv
mlflow
flask
autogluon
pandas
kagglehub
confluent_kafka